Add OrdersToProductsRepository + fix some entities bugs

+++ SalesOut.DAL/Models/Entities/Implementations/OrderToProduct.cs
@@ -9,7 +9,7 @@ namespace SalesOut.DAL.Models.Entities.Implementations
 
         public ulong OrderId { get; set; }
 
-        public ulong Quantity { get; set; }
+        public int Quantity { get; set; }
 
         public override string ToString()
         {
